flip triangles in csg_tree
